feat: move nodepool ordered maestro readonly bundle deletion to create nodepool syncer
